Validate input before adding
This adds a simple check to make sure the user typed something into the input before hitting submit. If the 'folder-name' input is a non-empty string then itll proceed with `addFolder()`.

In the `else` block I just added some `state` to keep track of whether or not to show an error to the user. That error is being rendered on line 48.

Also on line 21 you see I'm clearing the `folderNameError` state so that when user tries to submit after getting an error the error goes away.